data-binding with recyclerview

+++ app/src/main/java/com/balajiprabhu/weatherapp/network/model/MainValues.java
@@ -8,9 +8,9 @@ public class MainValues {
     @SerializedName("temp")
     private Double temp;
     @SerializedName("pressure")
-    private Integer pressure;
+    private Double pressure;
     @SerializedName("humidity")
-    private Integer humidity;
+    private Double humidity;
     @SerializedName("temp_min")
     private Double tempMin;
     @SerializedName("temp_max")
@@ -24,19 +24,19 @@ public class MainValues {
         this.temp = temp;
     }
 
-    public Integer getPressure() {
+    public Double getPressure() {
         return pressure;
     }
 
-    public void setPressure(Integer pressure) {
+    public void setPressure(Double pressure) {
         this.pressure = pressure;
     }
 
-    public Integer getHumidity() {
+    public Double getHumidity() {
         return humidity;
     }
 
-    public void setHumidity(Integer humidity) {
+    public void setHumidity(Double humidity) {
         this.humidity = humidity;
     }
